Added missing book retrieval endpoint and correction to code
