+++ api/v1/router.py
@@ -1,6 +1,6 @@
 from fastapi import APIRouter
 
-from api.routes import books
+from api.v1.routes import books
 
 api_router = APIRouter()
 api_router.include_router(books.router, prefix="/books", tags=["books"])
